+++ src/DataType/IDFilterFactory.php
@@ -7,19 +7,20 @@
 
 declare(strict_types=1);
 
-namespace OxidEsales\GraphQL\Base\DataObject;
+namespace OxidEsales\GraphQL\Base\DataType;
 
 use TheCodingMachine\GraphQLite\Annotations\Factory;
+use TheCodingMachine\GraphQLite\Types\ID;
 
-class BoolFilterFactory
+class IDFilterFactory
 {
     /**
      * @Factory()
      */
-    public static function createBoolFilter(
-        bool $equals
-    ): BoolFilter {
-        return new BoolFilter(
+    public static function createIDFilter(
+        ID $equals
+    ): IDFilter {
+        return new IDFilter(
             $equals
         );
     }