+++ tests/Integration/Framework/Controller/TestController.php
@@ -11,7 +11,7 @@ namespace OxidEsales\GraphQL\Base\Tests\Integration\Framework\Controller;
 
 use OxidEsales\GraphQL\Base\Exception\InvalidToken;
 use OxidEsales\GraphQL\Base\Exception\NotFound;
-use OxidEsales\GraphQL\Base\Tests\Integration\Framework\DataObject\TestFilter;
+use OxidEsales\GraphQL\Base\Tests\Integration\Framework\DataType\TestFilter;
 use TheCodingMachine\GraphQLite\Annotations\Query;
 use TheCodingMachine\GraphQLite\Annotations\Logged;
 use TheCodingMachine\GraphQLite\Annotations\Right;